requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.3
